requirements_lite.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
opencv-python-headless
pillow
